backend/requirements.txt
Django==5.2.1
djangorestframework==3.16
django-filter==25.1
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0